FIX: exit 0
The restricted security_opt blocked the removing of orphaned files by
matplotlib/streamlit. This resulted in an exit of the container, without
printing an error message, which made debugging difficult.

Changed tmp directories for streamlit an matplotlib and introduced a
dedicated user for the container. Now the container does not crash
anymore.